- Fix end_backup.sh script - Add logging in end_backup.sh

+++ root/app/end_backup.sh
@@ -1,4 +1,12 @@
 #!/usr/bin/with-contenv bash
-pid=$(ps | grep "duplicacy backup" | grep -v "grep" | cut -d " " -f 1)
-kill -2 $dup_pid
+
+dup_pid=`ps | grep "duplicacy backup" | grep -v "grep" | awk '{print $1;}'`
+
+if [[ ! -z ${dup_pid} ]]; then
+    echo Stop current backup process with safe resume from current uploaded progress when next backup occurs
+    kill -2 $dup_pid
+else
+    echo No current backup process, nothing to stop
+fi
+
 exit 0
